Import Excel - add check before deleting rows Reviewer: Daniel Jao

+++ src/main/java/org/generationcp/middleware/service/FieldbookServiceImpl.java
@@ -936,7 +936,9 @@ public class FieldbookServiceImpl extends Service implements FieldbookService {
 				}
 			}
 		
-			getExperimentDestroyer().deleteExperimentsByIds(deletedExperimentIds);
+			if (!deletedExperimentIds.isEmpty()) {
+				getExperimentDestroyer().deleteExperimentsByIds(deletedExperimentIds);
+			}
 		}
 	}
 }
